requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtplib==3.0.1
aiostream==0.5.2
altair==5.1.2
annotated-types==0.6.0
anthropic==0.28.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
asgiref==3.7.2
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.6.0
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
blessed==1.20.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chainlit==0.7.604
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
config==0.5.1
contourpy==1.2.0
courlan==0.9.4
coverage==7.5.4
crypto==1.4.1
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.5.14
dateparser==1.2.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distlib==0.3.7
distro==1.8.0
dnspython==2.4.2
docstring_parser==0.16
et-xmlfile==1.1.0
eval_type_backport==0.2.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.100.0
fastapi-socketio==0.0.10
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
flatbuffers==24.3.25
fonttools==4.45.1
frozendict==2.3.9
frozenlist==1.4.0
fsspec==2023.10.0
gast==0.5.4
geojson==3.1.0
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.6.4
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-aiplatform==1.56.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-pasta==0.2.0
google-resumable-media==2.7.1
google-search-results==2.4.2
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
gspread==6.1.2
gspread-dataframe==3.3.1
h11==0.14.0
h5py==3.11.0
html2image==2.0.4.3
html2text==2024.2.26
html5lib==1.1
htmldate==1.6.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.8.0
inquirer==3.1.4
ipykernel==6.26.0
ipython==8.18.0
ipywidgets==8.1.3
itsdangerous==2.1.2
jaraco.context==4.3.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.4.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab_widgets==3.0.11
jusText==3.0.0
keras==3.3.3
kiwisolver==1.4.5
langchain==0.0.340
langcodes==3.3.0
langsmith==0.0.66
Lazify==0.4.0
libclang==18.1.1
litellm==1.7.5
llama-index==0.9.8
loguru==0.7.2
lxml==4.9.2
mailchimp-transactional==1.0.50
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
mypy==1.10.1
mypy-extensions==1.0.0
Naked==0.1.32
namex==0.0.8
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.7
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
odrive==0.6.8
open-interpreter==0.1.16
openai==1.3.5
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.1.3
pandas-datareader==0.10.0
pandas-stubs==2.2.2.240603
parso==0.8.3
peewee==3.17.0
pexpect==4.9.0
phonenumbers==8.13.26
pikepdf==8.7.1
pillow==10.3.0
pinecone-client==2.2.4
pip-review==1.3.0
pkce==1.0.3
platformdirs==4.0.0
prisma==0.11.0
prompt-toolkit==3.0.41
promptlayer==0.4.0
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.8.0
pydub==0.25.1
pyelftools==0.31
Pygments==2.17.2
PyJWT==2.8.0
Pympler==1.0.1
pyOpenSSL==23.3.0
pyowm==3.3.0
pyparsing==3.1.1
pypdf==3.17.1
PyPDF2==3.0.1
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-editor==1.0.4
python-engineio==4.8.0
python-graphql-client==0.4.3
python-http-client==3.3.7
python-multipart==0.0.6
python-pptx==0.6.23
python-socketio==5.10.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pyusb==1.2.1
PyYAML==6.0.1
pyzmq==25.1.1
readchar==4.0.5
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3986==2.0.0
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.4
sendgrid==6.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.4
shellescape==3.8.1
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.27.0
streamlit==1.28.2
stripe==7.6.0
style==1.1.0
supertokens-python==0.17.0
sympy==1.12
syncer==2.0.3
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.16.1
tensorflow-metal==1.1.0
termcolor==2.3.0
textblob==0.17.1
threadpoolctl==3.2.0
tiktoken==0.5.1
tld==0.13
tldextract==5.1.1
together==1.2.0
tokenizers==0.15.0
tokentrim==0.1.13
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.4
trafilatura==1.6.2
traitlets==5.13.0
transformers==4.35.2
twilio==8.10.2
typer==0.12.3
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
update==0.0.1
uptrace==1.21.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.23.2
validators==0.22.0
virtualenv==20.24.7
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
wget==3.2
widgetsnbextension==4.0.11
wolframalpha==5.0.0
wordcloud==1.9.2
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.1.9
xmltodict==0.13.0
yarl==1.9.3
yaspin==3.0.1
yfinance==0.2.32
zipp==3.17.0
